Add cloud trace request sampling.
Set up a QPS based sampling mechanism based on the Cloud Trace team's
suggestions. The QPS can be configured in server config and have a default
value 0.1. Trace is triggered if the time interval between the request time
and the previous trace enabled request is bigger than a threshold. The
threshold is calculated from the qps configuration. When a request has trace
context header turning on trace, the timestamp for the previous traced request
will be refreshed to the current time.

Change-Id: I16304fc0c785c4b5020cf4ae00c3f97b4a6d0931