Propogate error via status code when transcoding
When transcoding a REST client call to the gRPC backend, delay sending
the headers to the client until the first response message arrives from
from the backend or until the end of the call. This will allow us to
propogate the errors from the backend using the HTTP status code in the
case when the backend returns the error before the first message.

Plus, add error tests for transcoding for the following cases:
- service-conrtrol check fails,
- backend unavailable case,
- error propagated from the backend,
- method does not exist.

Change-Id: Ib3f2c9aab9a0be05d7fd0f5666a7d0771ec7c753