Respect allow_cors configuration.
CORS is enabled only if "allow_cors" configuration in swagger config is
set to true. By default, ESP does not allow CORS traffic, which means
that the default behavior of ESP is to reject all "OPTIONS" requests.
If customers want to enable CORS, they need to set "allow_cors" to true
in swagger config.

Note that this is differnet from the default behavior of ESP Beta, which
is to allow all CORS traffic.

Change-Id: I464ce785a328ce9687c32e8a6cbe47a8aef2359f