Accept Issuer with or without https prefix for OpenID discovery.
Currently for OpenID discovery, we require that the issuer in JWT
matches exactly the issuer listed in service config. For example, if
"accounts.google.com" is listed as an issuer in service config, and a
token has issuer "https://accounts.google.com", ESP will fail to
validate the token because OpenID discovery cannot find the matching
issuer for "https://accounts.google.com". This change fixes the problem.

Change-Id: Ic173b7ce600f0e976de5555d122d4421de235443