Fix memory leak caused by aborted client calls
When writing to the output socket fails (e.g. if the client has died),
ngx_http_output_filter() returns NGX_ERROR. Then, ESP passes it to the
ngx_http_finalize_request(), which in the case of NGX_ERROR leaks the
request.
The fix is to convert NGX_ERROR to NGX_DONE before passing the status to
ngx_http_finalize_request().

Change-Id: I185fac2cededd278d0b8c3ba6b25db9b8e0e492c