Fix a shutdown crash found by TSAN tests
The crash happens when there are still pending callbacks in the gRPC
queue when shutting down. This happens when the request has been
finalized due to an error, but there was still a call to the backend in
flight. As there are no more requests NGINX thinks it's safe to
shutdown. Processing these events try to enqueue new events into the
gRPC completion queue, which causes the crash as the completion queue
has been shut down.

Change-Id: I210be89b1e59c6893a99989171e63c36a6f08135