Feature: TLS for gRPC
This changeset fixes and closes cloudendpoints/esp#346, by adding
the ability to invoke the gRPC channel constructor with SSL/TLS
credentials based on nginx config.

Basically, a few config entries are added, which are interpreted
at startup and then passed to the gRPC `SslCredentials` struct:

New config file entries:
- `grpc_roots`: Specify a file full of PEMs to use as authorities,
   when operating from the client's perspective (so, authorities
   for upstream gRPC backends).

- `grpc_client_key`: Kind of obvious. Lets you specify a PEM file
  to load and use as a client-side private key.

- `grpc_client_certificte`: Matches the `grpc_client_key` with a
  path to a PEM file containing a client-side certificate.

Behavior:
If either `grpc_client_key` or `grpc_client_certificate` is
specified, the other must be specified. Both are loaded and kept
for when the gRPC channel is initialized.

If `grpc_roots` is 1) left unspecified, or 2) specified with the
special value `default`, the default root authorities from gRPC
are used.